realtek-poe: move protocol-agnostic code to common header
Move the definitions and data structures which do not deal with a
specific MCU protocol to a separate header. main.c is getting quite
long and convoluted, so let's try to ease the congestion.
